Flaky repro for compaction issue
ossl_ssl_initialize() sets a Ruby reference on an OpenSSL handle and
it's retrived later for calling callbacks such as session_new_cb. There
is nothing marking the reference that lives on the SSL handle, so the
reference could move and cause crashes later when the stale reference is
used.

On my machine I get a crash with this change running the following tests
in a loop:

    bundle exec rake test TEST=test/openssl/test_ssl_session.rb TESTOPTS='-n/cb/'